refactor(tui): P12.2 extract InputAction dispatch into dispatch.rs (LoopState)
The ~1200-line match moved behind dispatch::handle_input_action(action, LoopState<'_>)
-> ControlFlow. LoopState bundles the loop's &mut borrows (app/runtime/config/stream/
handles/registry/...). No escape hatch needed — borrow bundle held. mod.rs 2487->~1160.
Behavior byte-identical: 331 unit + 38 harness + smoke green, clippy 0.
SMOKE CHECKLIST for review: submit, slash cmd, abort, settings-apply, plugin-install, gamba.
